tests/wpt/mozilla/tests/mozilla/mozbrowser/iframe_goback.html

<head>
<title>iframe.goBack()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {

  var url2 = "data:,a";
  var url1 = `data:text/html,<script>setTimeout(() => location.assign("${url2}"), 0)</${"script"}>`;

  var locations = []
  var expected_locations = [url1, url2, url1];

  var iframe = document.createElement("iframe");
  iframe.mozbrowser = "true";
  iframe.src = url1;

  iframe.addEventListener("mozbrowserlocationchange", e => {
    locations.push(e.detail.uri);
    if (e.detail.uri == url2) {
      iframe.goBack();
    }
    if (locations.length == expected_locations.length) {
       assert_array_equals(locations, expected_locations);
       t.done();
    }
  });

  document.body.appendChild(iframe);

});

</script>
</body>
